# faster spd checking