--- a/geomstats/_backend/__init__.py
+++ b/geomstats/_backend/__init__.py
@@ -132,6 +132,7 @@ BACKEND_ATTRIBUTES = {
         "eigvalsh",
         "expm",
         "inv",
+        "is_pd",
         "logm",
         "norm",
         "qr",