--- a/geomstats/_backend/autograd/linalg.py
+++ b/geomstats/_backend/autograd/linalg.py
@@ -103,10 +103,12 @@ def qr(*args, **kwargs):
     )(*args, **kwargs)
 
 
-def _is_single_matrix_pd(mat):
+def is_single_matrix_pd(mat):
     """Check if a two dimensional square matrix is
     positive definite
     """
+    if mat.dim[0] != mat.dim[1]:
+        return False
     try:
         np.linalg.cholesky(mat)
         return True
@@ -114,16 +116,3 @@ def _is_single_matrix_pd(mat):
         if e.args[0] == "Matrix is not positive definite":
             return False
         raise e
-
-
-def is_pd(mat):
-    """Check if matrix is positive definite matrix
-    (doesn't check if its symmetric)
-    """
-    if mat.ndim == 2 and mat.shape[0] == mat.shape[1]:
-        return np.asarray(_is_single_matrix_pd(mat))
-    if mat.ndim == 2 and mat.shape[0] != mat.shape[1]:
-        return np.asarray(False)
-    if mat.ndim == 3 and mat.shape[1] == mat.shape[2]:
-        return np.asarray([_is_single_matrix_pd(m) for m in mat])
-    return np.asarray([False] * mat.shape[0])