--- a/geomstats/_backend/numpy/linalg.py
+++ b/geomstats/_backend/numpy/linalg.py
@@ -95,6 +95,6 @@ def is_pd(mat):
     elif mat.ndim == 2 and mat.shape[0] != mat.shape[1]:
         return False
     elif mat.ndim == 3 and mat.shape[1] == mat.shape[2]:
-        return [_is_single_matrix_pd(m) for m in mat]
+        return np.asarray([_is_single_matrix_pd(m) for m in mat])
     elif mat.ndim == 3 and mat.shape[1] != mat.shape[2]:
-        return [False] * mat.shape[0]
+        return np.asarray([False] * mat.shape[0])