--- a/geomstats/_backend/numpy/linalg.py
+++ b/geomstats/_backend/numpy/linalg.py
@@ -91,10 +91,10 @@ def is_pd(mat):
     (doesn't check if its symmetric)
     """
     if mat.ndim == 2 and mat.shape[0] == mat.shape[1]:
-        return _is_single_matrix_pd(mat)
-    elif mat.ndim == 2 and mat.shape[0] != mat.shape[1]:
-        return False
-    elif mat.ndim == 3 and mat.shape[1] == mat.shape[2]:
+        return np.asrray(_is_single_matrix_pd(mat))
+    if mat.ndim == 2 and mat.shape[0] != mat.shape[1]:
+        return np.asarray(False)
+    if mat.ndim == 3 and mat.shape[1] == mat.shape[2]:
         return np.asarray([_is_single_matrix_pd(m) for m in mat])
-    elif mat.ndim == 3 and mat.shape[1] != mat.shape[2]:
+    if mat.ndim == 3 and mat.shape[1] != mat.shape[2]:
         return np.asarray([False] * mat.shape[0])