--- a/geomstats/_backend/numpy/linalg.py
+++ b/geomstats/_backend/numpy/linalg.py
@@ -75,7 +75,7 @@ def qr(*args, **kwargs):
 
 def is_single_matrix_pd(mat):
     """Check if a two dimensional square matrix is
-    positive definite
+    positive definite.
     """
     if mat.shape[0] != mat.shape[1]:
         return False