--- a/geomstats/_backend/pytorch/linalg.py
+++ b/geomstats/_backend/pytorch/linalg.py
@@ -77,19 +77,16 @@ def qr(*args, **kwargs):
     return tensor_q, tensor_r
 
 
+# (TODO) (sait) torch.linalg.cholesky_ex for even faster way   
 def _is_single_matrix_pd(mat):
     """Check if a two dimensional square matrix is
     positive definite
     """
     try:
-        ch = torch.linalg.cholesky(m)
+        ch = torch.linalg.cholesky(mat)
         return True
-    except np.linalg.LinAlgErro as e:
-        if e.args[0] == "Matrix is not positive definite":
-            return False
-        else:
-            raise e
-
+    except RuntimeError as e:
+        return False
 
 def is_pd(mat):
     """Check if matrix is positive definite matrix