--- a/geomstats/_backend/pytorch/linalg.py
+++ b/geomstats/_backend/pytorch/linalg.py
@@ -83,7 +83,7 @@ def _is_single_matrix_pd(mat):
     positive definite
     """
     try:
-        ch = torch.linalg.cholesky(mat)
+        torch.linalg.cholesky(mat)
         return True
     except RuntimeError as e:
         return False
@@ -98,6 +98,6 @@ def is_pd(mat):
     elif mat.ndim == 2 and mat.shape[0] != mat.shape[1]:
         return False
     elif mat.ndim == 3 and mat.shape[1] == mat.shape[2]:
-        return [_is_single_matrix_pd(m) for m in mat]
+        return torch.BoolTensor([_is_single_matrix_pd(m) for m in mat])
     elif mat.ndim == 3 and mat.shape[1] != mat.shape[2]:
-        return [False] * mat.shape[0]
+        return torch.BoolTensor([False] * mat.shape[0])