--- a/geomstats/_backend/pytorch/linalg.py
+++ b/geomstats/_backend/pytorch/linalg.py
@@ -85,7 +85,7 @@ def _is_single_matrix_pd(mat):
     try:
         torch.linalg.cholesky(mat)
         return True
-    except RuntimeError as e:
+    except RuntimeError as _e:
         return False
 
 
@@ -95,9 +95,9 @@ def is_pd(mat):
     """
     if mat.ndim == 2 and mat.shape[0] == mat.shape[1]:
         return torch.tensor(_is_single_matrix_pd(mat))
-    elif mat.ndim == 2 and mat.shape[0] != mat.shape[1]:
+    if mat.ndim == 2 and mat.shape[0] != mat.shape[1]:
         return torch.tensor(False)
-    elif mat.ndim == 3 and mat.shape[1] == mat.shape[2]:
+    if mat.ndim == 3 and mat.shape[1] == mat.shape[2]:
         return torch.tensor([_is_single_matrix_pd(m) for m in mat])
-    elif mat.ndim == 3 and mat.shape[1] != mat.shape[2]:
+    if mat.ndim == 3 and mat.shape[1] != mat.shape[2]:
         return torch.tensor([False] * mat.shape[0])