--- a/geomstats/_backend/tensorflow/linalg.py
+++ b/geomstats/_backend/tensorflow/linalg.py
@@ -87,9 +87,9 @@ def is_pd(mat):
     (doesn't check if its symmetric)
     """
     if mat.ndim == 2 and mat.shape[0] == mat.shape[1]:
-        return _is_single_matrix_pd(mat)
+        return tf.convert_to_tensor(_is_single_matrix_pd(mat))
     if mat.ndim == 2 and mat.shape[0] != mat.shape[1]:
-        return False
+        return tf.convert_to_tensor(False)
     if mat.ndim == 3 and mat.shape[1] == mat.shape[2]:
         return tf.convert_to_tensor([_is_single_matrix_pd(m) for m in mat])
     if mat.ndim == 3 and mat.shape[1] != mat.shape[2]: