… geomstats/geometry/spd_matrices.py
… tests/tests_geomstats/test_backends.py
… tests/tests_geomstats/test_matrices.py