faster SPD checking by cholesky decomposition
- [x] numpy
- [x] autograd
- [x] tensorflow
- [x] pytorch
- [x] tests

some simple benchmarks on numpy backend.
|size |  geomstats time| new time|
|------| -------------------- | -----  |  
| (10, 2000, 2000) | 21 sec | 3sec |
| (100 , 1000, 1000) | 25 sec | 4 sec|
|  (1000, 200, 200)  | 5 sec | ~0 sec |