@SaitejaUtpala 2021-12-15
@ninamiolane  I think we should keep _is_single_matrix_pd in the backends and move is_pd to the matrices module.. because currently I am using  torch.linalg.cholesky, but there is https://pytorch.org/docs/stable/generated/torch.linalg.cholesky_ex.html (still in experimental stage). but it can be even faster than torch.linalg.cholesky for pd checking ( quoting `This makes this function a faster way to check if a matrix is positive-definite, and it provides an opportunity to handle decomposition errors more gracefully or performantly than torch.linalg.cholesky() does.`)